rewrote tests for nav given changes with Body object
